Window/Image: Optimize DisplayImage()
Displaying an Utils::Image object on System::Window was optimized
on both platforms:
* On Windows creation of temporary HBITMAP object was removed.
  Instead, StretchDIBits function is called to directly draw contents
  of Image on Window.
* On Linux xcb_create_image_native() call was moved to Utils::Image
  and is called on Image creation. Because this call does a shallow
  copy of data pointer to xcb_image_t structure, xcb_image_put() call
  in Window can get updated contents without extra work, preventing
  useless allocations.

+++ lkCommonTest/Tests/Utils/PixelTest.cpp
@@ -50,7 +50,7 @@ const PixelFloat4 TEST_PIXEL_FLOAT_DIV_3_CONSTANT   ({  10.0f,  15.0f,  20.0f,  
 
 const float TEST_CONSTANT_CLAMP_FLOAT_POSITIVE = 2.0f;
 const float TEST_CONSTANT_CLAMP_FLOAT_NEGATIVE = -0.5f;
-const float TEST_CONSTANT_CLAMP_UINT = 255;
+const uint8_t TEST_CONSTANT_CLAMP_UINT = 255;
 
 
 TEST(Pixel, DefaultConstructor)
